waitForElement: new userscript library
Function `waitForElement` has been very useful in userscripts that I
develop and maintain.[1,2]  I've been copy-pasting it from script to
script, whenever it's been needed.

Originally, it was pretty much the verbatim code from Stack Overflow
answer by Yong Wang.[3]  However, over time, as I got more experience
with JavaScript, I tweaked it a little bit.[4,5]

To make usage of this function consistent across the userscripts [1,2]
publish it as a library for userscripts.  There is a similar library on
GitHub [6] that doesn't disconnect() the MutationObserver automatically.

In the future, I might add a `parent` parameter, which would have a
default value of `document` to change on which element function
querySelector gets called.

[1] https://github.com/search?q=owner%3Arybak+topic%3Auserscript&type=repositories
[2] https://greasyfork.org/en/users/951503-andrybak?language=js&sort=created
[3] https://stackoverflow.com/a/61511955/1083697
[4] 444f3b8 (waitForElement: disconnect observer before resolving,
    2023-08-19)
    https://github.com/rybak/copy-commit-reference-userscript/commit/444f3b8a99569085ae9495976c9884cddce6882b
[5] 9909b72 (waitForElement: call document.querySelector less,
    2023-08-19)
    https://github.com/rybak/copy-commit-reference-userscript/commit/9909b7221318405ac37da4ed5e4f2cd22924aeac
[6] https://github.com/alexanderjkslfj/queryObserver